Consolidate input vs output port-preview code
