+++ Stitch/Graph/Node/View/NodesView.swift
@@ -163,281 +163,14 @@ struct CanvasEdgesViewModifier: ViewModifier {
                                     document: document,
                                     graphUI: document.graphUI)
                 
-                
-                
+                // TODO: does PortPreviewPopoverView render too many times when open?
+                // TODO: more elegant way to do this? Generic types giving Swift compiler trouble
                 if let openPortPreview = graphUI.openPortPreview {
-                    
-                    // Find the input or output that has the matching canvas-item-id and row-observer-id
-                    
-//                    
-//                    if let rowViewModel = (openPortPreview.nodeIO == .input ? allInputs : allOutputs).first { rowViewModel in
-//                        rowViewModel.canvasItemDelegate?.id == openPortPreview.canvasItemId
-//                        && rowViewModel?.rowDelegate?.id == openPortPreview.port
-//                    } {
-//                        PortPreviewPopoverView(rowViewModel: rowViewModel,
-//                                               nodeIO: openPortPreview.nodeIO)
-//                    }
-//                    
-                    switch openPortPreview.nodeIO {
-                    case .input:
-                        if let rowViewModel = allInputs.first(where: { rowViewModel in
-                            rowViewModel.canvasItemDelegate?.id == openPortPreview.canvasItemId
-                            && rowViewModel.rowDelegate?.id == openPortPreview.port
-                        }) {
-                            PortPreviewPopoverView<InputNodeRowObserver>(rowViewModel: rowViewModel,
-                                                   nodeIO: .input)
-                        }
-//                        InputPreviewPopoverView(ports: allInputs,
-//                                                graphUI: graphUI)
-                    case .output:
-                        if let rowViewModel = allOutputs.first(where: { rowViewModel in
-                            rowViewModel.canvasItemDelegate?.id == openPortPreview.canvasItemId
-                            && rowViewModel.rowDelegate?.id == openPortPreview.port
-                        }) {
-                            PortPreviewPopoverView<OutputNodeRowObserver>(rowViewModel: rowViewModel,
-                                                   nodeIO: .output)
-                        }
-//                        OutputPreviewPopoverView(ports: allOutputs,
-//                                                 graphUI: graphUI)
-                    }
+                    PortPreviewPopoverWrapperView(
+                        allInputs: allInputs,
+                        allOutputs: allOutputs,
+                        openPortPreview: openPortPreview)
                 }
             }
-    }
-}
-
-
-let PORT_PREVIEW_POPOVER_MAX_HEIGHT: CGFloat = 420
-
-////struct PortPreviewPopoverView<NodeRowObserverType: NodeRowObserver>: View {
-//struct InputPreviewPopoverView: View {
-//
-//    let ports: [InputNodeRowViewModel]
-//    @Bindable var graphUI: GraphUIState
-//    
-//    @State private var width: CGFloat = .zero
-//    
-//    var body: some View {
-//        
-//        // If we have an open port preview
-//        if let openPortPreview = graphUI.openPortPreview,
-//           openPortPreview.nodeIO == .input {
-////        if case let .input(portPreviewInputObserverCoordinate, portPreviewCanvasItemId) = graphUI.openPortPreview {
-//            ForEach(ports) { inputRowViewModel in
-//                let portPreviewCanvasItemId = openPortPreview.canvasItemId
-//                
-//                 if inputRowViewModel.canvasItemDelegate?.id == openPortPreview.canvasItemId,
-//                    let rowObserver = inputRowViewModel.rowDelegate,
-//                    rowObserver.id == openPortPreview.port,
-//                    let anchor = inputRowViewModel.anchorPoint {
-//                    
-//                     
-//                     ZStack {
-//                         Rectangle().fill(.clear)
-//                             .frame(width: 30, height: 30)
-//                             .background(.ultraThinMaterial)
-// //                            .background(.red)
-//                             .rotationEffect(.degrees(45))
-//                             .position(x: anchor.x - self.width/2,
-//                                       y: anchor.y)
-//                             .offset(x: self.width/2 - 36)
-//                                                 
-//                         PortValuesPreviewView(
-//                            rowObserver: rowObserver,
-//                            rowViewModel: inputRowViewModel,
-//                            nodeIO: openPortPreview.nodeIO)
-//                         
-//                         .background {
-//                             GeometryReader { proxy in
-//                                 Color.clear
-//                                 // IMPORTANT: use .local frame, since .global is affected by zooming and creates infinite loop
-//                                     .onChange(of: proxy.frame(in: .local), initial: true) { _, newFrameData in
-//                                         log("InputPreviewPopoverView: newFrameData.size.width: \(newFrameData.size.width)")
-//                                         self.width = newFrameData.size.width
-//                                     }
-//                             }
-//                         }
-//                         .frame(maxHeight: PORT_PREVIEW_POPOVER_MAX_HEIGHT)
-// //                        .fixedSize(horizontal: true, vertical: true)
-//                         .fixedSize(horizontal: false, vertical: true)
-//                         
-//                         // self.width/1.65 is actually not a consistent
-//                         .position(x: anchor.x - self.width/2,
-//                                   y: anchor.y)
-//                         .offset(x: -32)
-//                         
-//                     } // ZStack
-//                     
-//                }
-//            }
-//        }
-//    }
-//}
-//
-//struct OutputPreviewPopoverView: View {
-//
-//    let ports: [OutputNodeRowViewModel]
-//    @Bindable var graphUI: GraphUIState
-//    
-//    @State private var width: CGFloat = .zero
-//    
-//    var body: some View {
-//        
-//        // If we have an open port preview
-//        if let openPortPreview = graphUI.openPortPreview,
-//           openPortPreview.nodeIO == .output {
-//            ForEach(ports) { outputRowViewModel in
-//                let portPreviewCanvasItemId = openPortPreview.canvasItemId
-//                
-//                 if outputRowViewModel.canvasItemDelegate?.id == openPortPreview.canvasItemId,
-//                    let rowObserver = outputRowViewModel.rowDelegate,
-//                    rowObserver.id == openPortPreview.port,
-//                    let anchor = outputRowViewModel.anchorPoint {
-//                    
-//                     ZStack {
-//                         Rectangle().fill(.clear)
-//                             .frame(width: 30, height: 30)
-//                             .background(.ultraThinMaterial)
-//                             .rotationEffect(.degrees(45))
-//                             .position(x: anchor.x + self.width/2,
-//                                       y: anchor.y)
-//                             .offset(x: -self.width/2 + 36)
-//                                                 
-//                         PortValuesPreviewView(
-//                            rowObserver: rowObserver,
-//                            rowViewModel: outputRowViewModel,
-//                            nodeIO: openPortPreview.nodeIO)
-//                         
-//                         .background {
-//                             GeometryReader { proxy in
-//                                 Color.clear
-//                                 // IMPORTANT: use .local frame, since .global is affected by zooming and creates infinite loop
-//                                     .onChange(of: proxy.frame(in: .local), initial: true) { _, newFrameData in
-//                                         log("OutputPreviewPopoverView: newFrameData.size.width: \(newFrameData.size.width)")
-//                                         self.width = newFrameData.size.width
-//                                     }
-//                             }
-//                         }
-//                         .frame(maxHeight: PORT_PREVIEW_POPOVER_MAX_HEIGHT)
-//                         .fixedSize(horizontal: false, vertical: true)
-//                         .position(x: anchor.x + self.width/2,
-//                                   y: anchor.y)
-//                         .offset(x: 32)
-//                         
-//                     } // ZStack
-//                     
-//                }
-//            }
-//        }
-//    }
-//}
-
-struct PortPreviewPopoverView<NodeRowObserverType: NodeRowObserver>: View {
-
-    let rowViewModel: NodeRowObserverType.RowViewModelType
-    let nodeIO: NodeIO
-    
-    @State private var width: CGFloat = .zero
-    
-    
-    func stablePopoverArrow(anchor: CGPoint,
-                            positionAdjustment: CGFloat) -> some View {
-        
-        let arrowOffsetAdjustment: CGFloat = nodeIO == .input ? (self.width/2 - 36) : (-self.width/2 + 36)
-        
-        return Rectangle().fill(.clear)
-            .frame(width: 30, height: 30)
-            .background(.ultraThinMaterial)
-            .rotationEffect(.degrees(45))
-            .position(x: anchor.x + positionAdjustment,
-                      y: anchor.y)
-            .offset(x: arrowOffsetAdjustment)
-    }
-    
-    func stablePopoverPreview(rowObserver: NodeRowObserverType) -> some View {
-        PortValuesPreviewView(
-            rowObserver: rowObserver,
-            rowViewModel: self.rowViewModel,
-            nodeIO: self.nodeIO)
-    }
-    
-    
-    var body: some View {
-        if let rowObserver: NodeRowObserverType = rowViewModel.rowDelegate,
-           let anchor = rowViewModel.anchorPoint {
-            
-            let positionAdjustment: CGFloat = nodeIO == .input ? -self.width/2 : self.width/2
-            let popoverOffsetAdjustment: CGFloat = nodeIO == .input ? -32 : 32
-            
-            ZStack {
-//                Rectangle().fill(.clear)
-//                    .frame(width: 30, height: 30)
-//                    .background(.ultraThinMaterial)
-//                    .rotationEffect(.degrees(45))
-//                    .position(x: anchor.x + positionAdjustment,
-//                              y: anchor.y)
-//                    .offset(x: arrowOffsetAdjustment)
-                
-                stablePopoverArrow(anchor: anchor,
-                                   positionAdjustment: positionAdjustment)
-                
-//                PortValuesPreviewView(
-//                    rowObserver: rowObserver,
-//                    rowViewModel: rowViewModel,
-//                    nodeIO: nodeIO)
-                
-                stablePopoverPreview(rowObserver: rowObserver)
-
-//                .background {
-//                    GeometryReader { proxy in
-//                        Color.clear
-//                        // IMPORTANT: use .local frame, since .global is affected by zooming and creates infinite loop
-//                            .onChange(of: proxy.frame(in: .local), initial: true) { _, newFrameData in
-//                                log("PortPreviewPopoverView: newFrameData.size.width: \(newFrameData.size.width)")
-//                                self.width = newFrameData.size.width
-//                            }
-//                    }
-//                }
-//                .frame(maxHeight: PORT_PREVIEW_POPOVER_MAX_HEIGHT)
-//                .fixedSize(horizontal: false, vertical: true)
-//                .position(x: anchor.x + positionAdjustment,
-//                          y: anchor.y)
-//                .offset(x: popoverOffsetAdjustment)
-                
-            } // ZStack
-            
-        }
-    }
-}
-
-
-
-struct EdgeInputLabelsView: View {
-    let inputs: [InputNodeRowViewModel]
-    @Bindable var document: StitchDocumentViewModel
-    @Bindable var graphUI: GraphUIState
-
-    var body: some View {
-        let showLabels = document.graphUI.edgeEditingState?.labelsShown ?? false
-        
-        if let nearbyCanvasItem: CanvasItemId = document.graphUI.edgeEditingState?.nearbyCanvasItem {
-            ForEach(inputs) { inputRowViewModel in
-                
-                // Doesn't seem to be needed? Checking the canvasItemDelegate seems to work well
-                // visibleNodeId property checks for group splitter inputs
-                // let isInputForNearbyNode = inputRowViewModel.visibleNodeIds.contains(nearbyCanvasItem)
-                
-                let isInputOnNearbyCanvasItem = inputRowViewModel.canvasItemDelegate?.id == nearbyCanvasItem
-                let isVisible = isInputOnNearbyCanvasItem && showLabels
-                
-                EdgeEditModeLabelsView(document: document,
-                                       portId: inputRowViewModel.id.portId)
-                .position(inputRowViewModel.anchorPoint ?? .zero)
-                .opacity(isVisible ? 1 : 0)
-                .animation(.linear(duration: .EDGE_EDIT_MODE_NODE_UI_ELEMENT_ANIMATION_LENGTH),
-                           value: isVisible)
-            }
-        } else {
-            EmptyView()
-        }
     }
 }
